Explicitly whitelist what we need
By default there is a ton of code being generated, including a bunch of
system types which aren't exposed in the MySQL API. For example, MySQL
has `#include <pthread.h>` somewhere. Bindgen is choking on the
definition of `pthread_mutex_t` on Travis, which is probably a bug. But
that type also never appears in any mysql type or function signature, so
we can just skip it.