Improve linking on windows
So there's a couple of things at play here. The first problem is that
`mysql_config` is completely useless on Windows, and `pkg-config` can't
find MySQL on Windows either. So we need to provide an override
directory.

Additionally, in their infinite wisdom, MySQL provides two separate
libraries, one for static linking and one for dynamic linking. The
static one requires the msvc toolchain. The dynamic one fails when run
with the latest msvc environment.