Update binding matching for MariaDB
The MariaDB bindings were not fully correct and missed some versions.
Also, the version indicated as 3.2 should have been 3.1, and 3.2 seems
to be compatible with 3.3 and thus should use those bindings.

I also found some errors in lib.rs where `cfg` matches were done twice
on 3.3, where i think one of those should be matching 3.4 also.

For the libmariadb version matching I'm using semver to keep it simple.
For example lib 10.6.0 should use 3.1, while 10.6.1 should use 3.3.
In the future there might be others of these minor differences and
semver can help make the matching easier.

I also added a contrib directory for the lak of currently knowing a
better place for it. This contains some scripts which are able to output
a mapping of which MariaDB Server uses which MariaDB Connector/C.

Fixes #65

Signed-off-by: BlackDex <black.dex@gmail.com>