Reenable the native_password login
This commit updates the mysql submodule to revert the commit that disabled the native_password authentication by default.
We now document which patches we apply on top of mysql to make future
development easier.

This commit also updates the bundled source code to mysql 9.3.0 and adds
bindings for that version.

Finally this commit contains a small fix for the build script output in
case of no matching version is found for the provided version by
exposing a clearer error message.