Add bundled_with_system_openssl feature
This commit adds a new feature `bundled_with_system_openssl` which causes the MySQL to use the system provided openssl library to link with.

Using this feature will prevent embedding an OpenSSL version and also dynamically link to an OpenSSL used by the application it self, not vendored.

Also updated the CI to match the CI's of pq-sys and Diesel.
- Removed MacOS v13 and replaced it with MacOS v15 Intel based
- Removed Windows 2022
- Updated the Ubuntu ARM build to 24.04

Signed-off-by: BlackDex <black.dex@gmail.com>